requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
PyPDF2
pinecone
sqlalchemy
psycopg2-binary
requests
python-multipart
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]